mythcommflag: rename commandlineparser.* to mythcommflag_...
