+++ mythtv/programs/mythcommflag/mythcommflag_commandlineparser.cpp
@@ -1,7 +1,7 @@
 #include <QString>
 
 #include "mythcorecontext.h"
-#include "commandlineparser.h"
+#include "mythcommflag_commandlineparser.h"
 
 MythCommFlagCommandLineParser::MythCommFlagCommandLineParser() :
     MythCommandLineParser(MYTH_APPNAME_MYTHCOMMFLAG)
